requirements.txt
flask
mysql-connector
gevent
sqlalchemy